+++ Tags.lua
@@ -58,7 +58,7 @@ function tags:setSortedTags()
 		tinsert(self.sortedTags, tag)
 	end
 	sort(self.sortedTags, function(tag1, tag2) return filterTags[tag1][1] < filterTags[tag2][1] end)
-	for i, tag in pairs(self.sortedTags) do
+	for i, tag in ipairs(self.sortedTags) do
 		filterTags[tag][1] = i
 	end
 end
@@ -84,7 +84,7 @@ function tags:mountOptionsMenu_Init(level)
 	local mountIndex, mountID = MountJournal.menuMountIndex, MountJournal.menuMountID
 
 	if level == 1 then
-		local active = select(4, C_MountJournal.GetMountInfoByID(mountID))
+		local _,_,_, active, _,_, isFavorite = C_MountJournal.GetMountInfoByID(mountID)
 		local needsFanfare = C_MountJournal.NeedsFanfare(mountID)
 		info.notCheckable = true
 
@@ -107,7 +107,8 @@ function tags:mountOptionsMenu_Init(level)
 		UIDropDownMenu_AddButton(info, level)
 
 		if not needsFanfare then
-			local isFavorite, canFavorite = C_MountJournal.GetIsFavorite(mountIndex)
+			local _, canFavorite = C_MountJournal.GetIsFavorite(mountIndex)
+			info.disabled = not canFavorite
 
 			if isFavorite then
 				info.text = BATTLE_PET_UNFAVORITE
@@ -120,15 +121,9 @@ function tags:mountOptionsMenu_Init(level)
 					C_MountJournal.SetIsFavorite(mountIndex, true)
 				end
 			end
-
-			if canFavorite then
-				info.disabled = false
-			else
-				info.disabled = true
-			end
 			UIDropDownMenu_AddButton(info, level)
 
-			info.disabled = false
+			info.disabled = nil
 			info.keepShownOnClick = true
 			info.hasArrow = true
 			info.func = nil
@@ -136,7 +131,7 @@ function tags:mountOptionsMenu_Init(level)
 			UIDropDownMenu_AddButton(info, level)
 		end
 
-		info.disabled = false
+		info.disabled = nil
 		info.keepShownOnClick = nil
 		info.hasArrow = nil
 		info.func = nil
@@ -197,7 +192,7 @@ function tags:addTag()
 		end
 		self.filter.tags[text] = {#self.sortedTags + 1, true}
 		tinsert(self.sortedTags, text)
-		journal:mountsListFullUpdate()
+		journal:updateMountsList()
 	end)
 end
 
@@ -209,7 +204,7 @@ function tags:deleteTag(tag)
 		end
 		self.filter.tags[tag] = nil
 		self:setSortedTags()
-		journal:mountsListFullUpdate()
+		journal:updateMountsList()
 	end)
 end
 
@@ -259,8 +254,8 @@ function tags:getFilterMount(mountID)
 		local i = 0
 		for tag, value in next, filterTags do
 			if value[2] then
-				i = i + 1
 				if not mountTags[tag] then return false end
+				i = i + 1
 			end
 		end
 		return i > 0