+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-* Fixed loading profile by specialization.
+* Added filter "Hidden For Character" (the filter system has been redone).
+* Increased maximum zoom distance from camera when viewing mount.
+* Battle Pet Search now looks for matches in source text.
+* Minor fixes.